test: shared getBalance tests for both facades
Extract getBalance coverage from monolithic facade test
files into the shared test framework. Shared tests run
against both fullnode and wallet-service adapters via
describe.each.

- shared/get-balance.test.ts: zero balance, funded balance
- fullnode-specific/get-balance.test.ts: custom token,
  mandatory tokenUid, internal transfer
- service-specific/get-balance.test.ts: no-arg getBalance,
  not-ready error, skipped empty-wallet bugs
